stream/server/requirements.txt
vidgear
colormath
opencv-python
Pillow
numpy
cppyy
yt_dlp
websockets